Use the eigendecomposition caching mechanism in {forward,backward}alpha_timeevolution
Previously the adiabatic time evolution functions used explicit
eigendecompositions. This was nicely straightforward, but it denied me the
opportunity to cache the eigendecompositions (and use cached
eigendecompositions).

I change to use those mechanisms. This requires me to be clever, which is not
good: need internal checks (store current alpha) to make sure I'm on the right
track.