thermalizer graph doesn't need to be directed
LightGraphs connected_component no longer works on directed graphs. We don't
really care about the fact that the graph is directed---this keeps track of the
fact that upward jumps happen much slower than downward jumps, and whenever we
ask about connected component we just want to know what the possibilities
are---so I replace DiGraph by Graph.